Return total time as a double

+++ spark-common/src/main/java/me/lucko/spark/common/sampler/node/StackTraceNode.java
@@ -75,7 +75,7 @@ public final class StackTraceNode extends AbstractNode implements Comparable<Sta
 
     @Override
     public int compareTo(StackTraceNode that) {
-        int i = -Long.compare(this.getTotalTime(), that.getTotalTime());
+        int i = -Double.compare(this.getTotalTime(), that.getTotalTime());
         if (i != 0) {
             return i;
         }
